pyproject.toml
[project]
name = "daydayarxiv"
version = "0.1.0"
description = "Daily arXiv information website. Python part is an updator."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.1.3",
    "loguru>=0.7.3",
    "langfuse>=3.0.0",
    "openai>=1.66.3",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.2.1",
    "pytest-cov>=5.0.0",
    "python-dotenv>=1.0.1",
    "pyright>=1.1.407",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.23.6",
    "pytest-mock>=3.14.0",
    "rich>=13.7.1",
    "ruff>=0.14.2",
    "tenacity>=8.2.3",
    "typer>=0.12.3",
]

[project.scripts]
daydayarxiv = "daydayarxiv.cli:main"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/daydayarxiv"]

[tool.uv]
package = true

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q --cov=daydayarxiv --cov-report=term-missing --cov-fail-under=100"
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100
include = ["src/daydayarxiv/**/*.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["daydayarxiv"]

[tool.ruff.lint.per-file-ignores]
"src/daydayarxiv/prompts/*.py" = ["RUF001"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
reportMissingTypeStubs = "none"
venvPath = "."
venv = ".venv"
include = ["src/daydayarxiv"]
exclude = ["**/__pycache__"]
